Add a runner interface to set a good infrastructure for upcoming changes to due an addition runner for vs
