+++ ftp/testfile
@@ -1,0 +1,1 @@
+Test data for decompression
