+++ runner/runner.go
@@ -1,61 +1,24 @@
 package runner
 
 import (
-	"context"
 	"errors"
-	"time"
-
-	"github.com/Kjone1/imageElevator/config"
-	"github.com/Kjone1/imageElevator/docker"
-	"github.com/Kjone1/imageElevator/ftp"
 	"github.com/rs/zerolog/log"
 )
 
-type Runner struct {
-	ctx             context.Context
-	sampleRate      time.Duration
-	timer           *time.Timer
-	runUploadChan   chan interface{}
-	resetTimerChan  chan interface{}
-	registryAdapter docker.RegistryAdapter
+type Runner interface {
+	uploadImages() (int, error)
+	pullFiles() ([]string, error)
+	Start()
+	TriggerUpload() error
 }
 
-func NewRunner(ctx context.Context, registryAdapter docker.RegistryAdapter, runnerConfig *config.RunnerConfiguration, registryConfig *config.RegistryConfiguration) *Runner {
-	timer := time.NewTimer(runnerConfig.SampleRateInMinutes)
-	runUploadChan := make(chan interface{}, 1)
-	resetTimerChan := make(chan interface{})
-
-	runner := &Runner{
-		ctx:             ctx,
-		sampleRate:      runnerConfig.SampleRateInMinutes,
-		timer:           timer,
-		runUploadChan:   runUploadChan,
-		resetTimerChan:  resetTimerChan,
-		registryAdapter: registryAdapter,
-	}
-
-	go runner.timerRoutine()
-	go runner.uploaderRoutine()
-
-	return runner
-}
-
-func (r *Runner) TriggerUpload() error {
-	select {
-	case r.runUploadChan <- nil:
-		return nil
-	default:
-		return errors.New("too many requests")
-
-	}
-}
-
-func (r *Runner) uploaderRoutine() {
+func uploaderRoutine(r *DockerRunner) {
 	log.Debug().Msg("Image uploader routine started")
 	for {
 		select {
 		case <-r.runUploadChan:
-			if err := r.uploadImages(); err != nil {
+			imagesCount, err := r.uploadImages()
+			if err != nil {
 				//FIXME: This is stupid!!! make it not an error somehow
 				if err.Error() == "no new images were found" {
 					log.Debug().Msg(err.Error())
@@ -63,11 +26,18 @@ func (r *Runner) uploaderRoutine() {
 					log.Err(err).Msgf("Failed to upload images from Image Uploader")
 				}
 			}
+
+			if imagesCount > 0 {
+				log.Info().Msgf("Uploaded %d images", imagesCount)
+			} else {
+				log.Debug().Msg("No images uploaded")
+			}
+
 			r.resetTimerChan <- nil
 
 		case <-r.ctx.Done():
 			log.Debug().Msg("Closing Image Uploader")
-			if err := ftp.Close(); err != nil {
+			if err := r.ftpClient.Close(); err != nil {
 				log.Err(err).Msg("Failed to close connection")
 			}
 
@@ -78,52 +48,12 @@ func (r *Runner) uploaderRoutine() {
 	}
 }
 
-func (r *Runner) uploadImages() error {
-	tarFiles, err := pullTarFile()
-	if err != nil {
-		return err
+func (r *DockerRunner) TriggerUpload() error {
+	select {
+	case r.runUploadChan <- nil:
+		return nil
+	default:
+		return errors.New("too many requests")
+
 	}
-
-	images := tarsToImages(tarFiles)
-
-	for i := 0; i < len(images); i++ {
-		if err := r.registryAdapter.PushTar(r.ctx, &images[i]); err != nil {
-			log.Err(err).Msgf("failed to push %s to registry", images[i].TarPath)
-		}
-	}
-
-	return nil
-}
-
-// TODO: Make a function receives a list of tar files and returns a docker.Image (ImageName, Tag, TarPath) by regex
-func tarsToImages(tarFiles []string) []docker.Image {
-	_ = tarFiles
-	return nil
-}
-
-func pullTarFile() ([]string, error) {
-
-	client, err := ftp.Client()
-	if err != nil {
-		log.Err(err).Msg("Unable to create FTP client with error")
-		return nil, err
-	}
-
-	images, err := ftp.List(client)
-	if err != nil {
-		log.Err(err).Msg("Reading FTP directory failed with error")
-		return nil, err
-	}
-
-	if images == nil {
-		return nil, errors.New("no new images were found")
-	}
-
-	files, err := ftp.Pull(client, images)
-	if err != nil {
-		log.Err(err).Msg("Pulling files from FTP directory")
-		return nil, err
-	}
-
-	return files, nil
 }
